extraer vista de usuario de urls
