+++ back-end/django/enbalde/urls.py
@@ -7,8 +7,8 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework import status
-from .models import Usuario, Articulo, TipoArticulo, Carrito, Seleccion, Venta, Envio, Oferta
-from .serializers import UsuarioSerializer, ArticuloSerializer, TipoArticuloSerializer, CarritoSerializer, \
+from .models import Articulo, TipoArticulo, Carrito, Seleccion, Venta, Envio, Oferta
+from .serializers import ArticuloSerializer, TipoArticuloSerializer, CarritoSerializer, \
     SeleccionSerializer, VentaSerializer, OfertaSerializer, EnvioSerializer
 from .views.logout_views import LogoutView
 from .views.carrito_views import UnCarrito, Carritos
@@ -17,31 +17,7 @@ from .views.common import generar_nombre_unico
 from .views.contacto_views import ContactoView
 from .views.registracion_views import SignupView
 from .views.login_views import LoginView
-
-
-class UsuarioViewSet(viewsets.ModelViewSet):
-    permission_classes = [IsAuthenticated]
-    queryset = Usuario.objects.all()
-    serializer_class = UsuarioSerializer
-
-    def update(self, request, *args, **kwargs):
-        usuario = self.get_object()
-
-        nueva_direccion = request.data.get('direccion')
-        nuevo_email = request.data.get('email')
-        nueva_clave = request.data.get('clave')
-        nuevo_telefono = request.data.get('telefono')
-        nuevas_observaciones = request.data.get('observaciones')
-
-        usuario.direccion = nueva_direccion
-        usuario.email = nuevo_email
-        usuario.telefono = nuevo_telefono
-        usuario.observaciones = nuevas_observaciones
-        usuario.set_password(nueva_clave)
-        usuario.save()
-
-        serializer = self.get_serializer(usuario)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+from .views.usuario_views import UsuarioViewSet
 
 
 class ArticuloViewSet(viewsets.ModelViewSet):
